Move keyboard shortcuts and feedback sections from sidebar to homepage
Both sections now appear in a responsive two-column grid on the homepage
(single column on mobile, two columns on sm+), making them visible on all
screen sizes instead of only on desktop sidebar.
